pyproject.toml
[project]
name = "personal-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "langchain-google-genai>=2.1.9",
    "langchain-openai>=0.3.28",
    "langgraph>=0.6.3",
]
